% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{sat_is_tracing}
\alias{sat_is_tracing}
\title{Is this solver recording a proof?}
\usage{
sat_is_tracing(solver)
}
\arguments{
\item{solver}{A \code{zusat_solver}.}
}
\value{
A single logical.
}
\description{
Is this solver recording a proof?
}
\examples{
s <- sat_solver()
sat_is_tracing(s)
sat_trace_proof(s, tempfile())
sat_is_tracing(s)
}
