% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{sat_n_vars}
\alias{sat_n_vars}
\title{Number of variables known to the solver}
\usage{
sat_n_vars(solver)
}
\arguments{
\item{solver}{A \code{zusat_solver}.}
}
\value{
An integer scalar.
}
\description{
Number of variables known to the solver
}
