% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{sat_signature}
\alias{sat_signature}
\title{Version of the bundled CaDiCaL}
\usage{
sat_signature()
}
\value{
A single string, for example \code{"cadical-3.0.1"}.
}
\description{
Version of the bundled CaDiCaL
}
\examples{
sat_signature()
}
