% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zusat-package.R
\docType{package}
\name{zusat-package}
\alias{zusat}
\alias{zusat-package}
\title{zusat: R Bindings for the 'CaDiCaL' SAT Solver}
\description{
Solves Boolean satisfiability (SAT) problems using 'CaDiCaL', a modern conflict-driven clause learning solver. Supports incremental solving, solving under assumptions, and inspection of which assumptions caused unsatisfiability. The 'CaDiCaL' C++ sources are bundled, so no external solver installation is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedrobtz/zusat}
  \item Report bugs at \url{https://github.com/pedrobtz/zusat/issues}
}

}
\author{
\strong{Maintainer}: Pedro Batista \email{pedrobtz@gmail.com}

Authors:
\itemize{
  \item Pedro Batista \email{pedrobtz@gmail.com}
}

Other contributors:
\itemize{
  \item Armin Biere (Author of the bundled CaDiCaL solver) [copyright holder]
  \item Mathias Fleury (Author of the bundled CaDiCaL solver) [copyright holder]
  \item Katalin Fazekas (Author of the bundled CaDiCaL solver) [copyright holder]
  \item Nils Froleyks (Author of the bundled CaDiCaL solver) [copyright holder]
}

}
\keyword{internal}
